+++ src/mod/info/Tables.d
@@ -528,7 +528,7 @@ public class Table
 
             public void display ( Output output, size_t width, ref char[] content_buf, ref char[] spacing_buf )
             {
-                size_t layoutSink ( char[] s )
+                uint layoutSink ( char[] s )
                 {
                     content_buf ~= s;
                     return s.length;
